fix(m3): make bf16 inference/train loading safe and unify numpy conversion   - replace m3 embedder .cpu().numpy() paths with base _convert_to_numpy(...) for bf16-safe conversion   - add torch_dtype plumbing in m3 runner model loading (AutoModel/colbert_linear/sparse_linear) to keep dtype behavior consistent

+++ FlagEmbedding/finetune/embedder/encoder_only/m3/runner.py
@@ -76,11 +76,13 @@ class EncoderOnlyEmbedderM3Runner(AbsEmbedderRunner):
         )
         colbert_linear = torch.nn.Linear(
             in_features=model.config.hidden_size,
-            out_features=model.config.hidden_size if colbert_dim <= 0 else colbert_dim
+            out_features=model.config.hidden_size if colbert_dim <= 0 else colbert_dim,
+            dtype=torch_dtype,
         )
         sparse_linear = torch.nn.Linear(
             in_features=model.config.hidden_size,
-            out_features=1
+            out_features=1,
+            dtype=torch_dtype,
         )
 
         colbert_model_path = os.path.join(model_name_or_path, 'colbert_linear.pt')
